Add function to create FileTime from SystemTime
This requires that FileTime can handle negative Unix timestamps on
non-Windows OSes. To prevent breaking existing API functions, they now
cast from signed to unsigned integers, but this may cause
overflow-related issues.

Nanoseconds have been left unsigned as it's harder to reason about if
both seconds and nanos are signed, and different platforms have varying
support for signed nanoseconds. For example, st_mtimensec is unsigned,
while suseconds_t only specifies the range [-1, 1,000,000].

+++ src/redox.rs
@@ -40,14 +40,14 @@ fn set_file_times_redox(fd: usize, atime: FileTime, mtime: FileTime) -> io::Resu
 
 pub fn from_last_modification_time(meta: &fs::Metadata) -> FileTime {
     FileTime {
-        seconds: meta.mtime() as u64,
+        seconds: meta.mtime(),
         nanos: meta.mtime_nsec() as u32,
     }
 }
 
 pub fn from_last_access_time(meta: &fs::Metadata) -> FileTime {
     FileTime {
-        seconds: meta.atime() as u64,
+        seconds: meta.atime(),
         nanos: meta.atime_nsec() as u32,
     }
 }
